# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BenchmarkDefinitionPayload": {
        "properties": {
          "benchmark_definition_id": {
            "title": "Benchmark Definition Id",
            "type": "string"
          },
          "cadence": {
            "title": "Cadence",
            "type": "string"
          },
          "candidate_scenario_ids": {
            "items": {
              "type": "string"
            },
            "title": "Candidate Scenario Ids",
            "type": "array"
          },
          "created_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At Utc"
          },
          "default_duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Duration Seconds"
          },
          "default_evaluation_profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Evaluation Profile Name"
          },
          "default_project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Project Id"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "focus_metrics": {
            "items": {
              "type": "string"
            },
            "title": "Focus Metrics",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "planning_mode": {
            "title": "Planning Mode",
            "type": "string"
          },
          "project_ids": {
            "items": {
              "type": "string"
            },
            "title": "Project Ids",
            "type": "array"
          },
          "queue_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Note"
          },
          "report_shape": {
            "title": "Report Shape",
            "type": "string"
          },
          "supports_duration_seconds": {
            "title": "Supports Duration Seconds",
            "type": "boolean"
          },
          "updated_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At Utc"
          }
        },
        "required": [
          "benchmark_definition_id",
          "name",
          "description",
          "focus_metrics",
          "cadence",
          "report_shape",
          "project_ids",
          "default_project_id",
          "default_evaluation_profile_name",
          "planning_mode",
          "candidate_scenario_ids",
          "supports_duration_seconds",
          "default_duration_seconds",
          "queue_note",
          "created_at_utc",
          "updated_at_utc"
        ],
        "title": "BenchmarkDefinitionPayload",
        "type": "object"
      },
      "BenchmarkDefinitionResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkDefinitionPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "BenchmarkDefinitionResponse",
        "type": "object"
      },
...
      "BenchmarkTaskListPayload": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkTaskPayload"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "tasks"
        ],
        "title": "BenchmarkTaskListPayload",
        "type": "object"
      },
      "BenchmarkTaskListResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkTaskListPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "BenchmarkTaskListResponse",
        "type": "object"
      },
...
      "BenchmarkTaskPayload": {
        "properties": {
          "auto_start": {
            "title": "Auto Start",
            "type": "boolean"
          },
          "benchmark_definition_id": {
            "title": "Benchmark Definition Id",
            "type": "string"
          },
          "benchmark_name": {
            "title": "Benchmark Name",
            "type": "string"
          },
          "benchmark_task_id": {
            "title": "Benchmark Task Id",
            "type": "string"
          },
          "counts_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts By Status",
            "type": "object"
          },
          "created_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At Utc"
          },
          "dut_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dut Model"
          },
          "ended_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At Utc"
          },
          "evaluation_profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Profile Name"
          },
          "hil_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HilConfigPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "planned_run_count": {
            "title": "Planned Run Count",
            "type": "integer"
          },
          "planning_mode": {
            "title": "Planning Mode",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "requested_duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Duration Seconds"
          },
          "run_ids": {
            "items": {
              "type": "string"
            },
            "title": "Run Ids",
            "type": "array"
          },
          "scenario_matrix": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkTaskScenarioMatrixEntryPayload"
            },
            "title": "Scenario Matrix",
            "type": "array"
          },
          "selected_scenario_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Scenario Ids",
            "type": "array"
          },
          "started_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At Utc"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/BenchmarkTaskSummaryPayload"
          },
          "updated_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At Utc"
          }
        },
        "required": [
          "benchmark_task_id",
          "project_id",
          "project_name",
          "benchmark_definition_id",
          "benchmark_name",
          "status",
          "planned_run_count",
          "counts_by_status",
          "run_ids",
          "scenario_matrix",
          "planning_mode",
          "selected_scenario_ids",
          "requested_duration_seconds",
          "hil_config",
          "evaluation_profile_name",
          "auto_start",
          "summary",
          "created_at_utc",
          "updated_at_utc",
          "started_at_utc",
          "ended_at_utc"
        ],
        "title": "BenchmarkTaskPayload",
        "type": "object"
      },
      "BenchmarkTaskResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkTaskPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "BenchmarkTaskResponse",
        "type": "object"
      },
...
      "BenchmarkTaskScenarioMatrixEntryPayload": {
        "properties": {
          "display_map_name": {
            "title": "Display Map Name",
            "type": "string"
          },
          "environment_name": {
            "title": "Environment Name",
            "type": "string"
          },
          "environment_preset_id": {
            "title": "Environment Preset Id",
            "type": "string"
          },
          "execution_backend": {
            "default": "native",
            "title": "Execution Backend",
            "type": "string"
          },
          "requested_map_name": {
            "title": "Requested Map Name",
            "type": "string"
          },
          "requested_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Timeout Seconds"
          },
          "resolved_map_name": {
            "title": "Resolved Map Name",
            "type": "string"
          },
          "resolved_timeout_seconds": {
            "title": "Resolved Timeout Seconds",
            "type": "integer"
          },
          "scenario_display_name": {
            "title": "Scenario Display Name",
            "type": "string"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_name": {
            "title": "Scenario Name",
            "type": "string"
          },
          "sensor_profile_name": {
            "title": "Sensor Profile Name",
            "type": "string"
          }
        },
        "required": [
          "scenario_id",
          "scenario_name",
          "scenario_display_name",
          "requested_map_name",
          "resolved_map_name",
          "display_map_name",
          "environment_preset_id",
          "environment_name",
          "sensor_profile_name",
          "resolved_timeout_seconds"
        ],
        "title": "BenchmarkTaskScenarioMatrixEntryPayload",
        "type": "object"
      },
      "BenchmarkTaskScenarioMatrixItemPayload": {
        "properties": {
          "environment_preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment Preset Id"
          },
          "map_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map Name"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "sensor_profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Profile Name"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "scenario_id"
        ],
        "title": "BenchmarkTaskScenarioMatrixItemPayload",
        "type": "object"
      },
...
      "CreateBenchmarkTaskRequest": {
        "properties": {
          "auto_start": {
            "default": false,
            "title": "Auto Start",
            "type": "boolean"
          },
          "benchmark_definition_id": {
            "title": "Benchmark Definition Id",
            "type": "string"
          },
          "dut_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dut Model"
          },
          "evaluation_profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Profile Name"
          },
          "hil_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HilConfigPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "run_duration_seconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Duration Seconds"
          },
          "scenario_matrix": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkTaskScenarioMatrixItemPayload"
            },
            "title": "Scenario Matrix",
            "type": "array"
          },
          "selected_scenario_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Scenario Ids",
            "type": "array"
          }
        },
        "required": [
          "benchmark_definition_id"
        ],
        "title": "CreateBenchmarkTaskRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "错误代码",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "错误说明",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HilConfigPayload": {
        "properties": {
          "dut_input_mode": {
            "default": "uvc_camera",
            "title": "Dut Input Mode",
            "type": "string"
          },
          "gateway_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "绑定的网关 ID。对于仅拉起 Host/Pi sidecar 的本地演示链路可为空。",
            "title": "Gateway Id"
          },
          "mode": {
            "default": "camera_open_loop",
            "title": "Mode",
            "type": "string"
          },
          "result_ingest_mode": {
            "default": "http_push",
            "title": "Result Ingest Mode",
            "type": "string"
          },
          "video_source": {
            "default": "hdmi_x1301",
            "title": "Video Source",
            "type": "string"
          }
        },
        "title": "HilConfigPayload",
        "type": "object"
      },
...
      "RerunBenchmarkTaskRequest": {
        "properties": {
          "auto_start": {
            "default": true,
            "title": "Auto Start",
            "type": "boolean"
          }
        },
        "title": "RerunBenchmarkTaskRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SimChip Nexus 芯境智测平台提供 CARLA 场景仿真、硬件在环测试、设备接入和评测报告服务。\n\n- `/docs`：接口文档\n- `/`：平台展示页\n- `/ui`：Web 测评控制台",
    "title": "SimChip Nexus API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/benchmark-definitions/{benchmark_definition_id}": {
      "get": {
        "operationId": "get_benchmark_definition_benchmark_definitions__benchmark_definition_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_definition_id",
            "required": true,
            "schema": {
              "title": "Benchmark Definition Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询单个基准定义",
        "tags": [
          "基准任务"
        ]
      }
    },
    "/benchmark-tasks": {
      "get": {
        "operationId": "list_benchmark_tasks_benchmark_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkTaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询基准任务列表",
        "tags": [
          "基准任务"
        ]
      },
      "post": {
        "operationId": "create_benchmark_task_benchmark_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBenchmarkTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "创建基准任务",
        "tags": [
          "基准任务"
        ]
      }
    },
    "/benchmark-tasks/{benchmark_task_id}": {
      "get": {
        "operationId": "get_benchmark_task_benchmark_tasks__benchmark_task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_task_id",
            "required": true,
            "schema": {
              "title": "Benchmark Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询单个基准任务",
        "tags": [
          "基准任务"
        ]
      }
    },
    "/benchmark-tasks/{benchmark_task_id}/rerun": {
      "post": {
        "operationId": "rerun_benchmark_task_benchmark_tasks__benchmark_task_id__rerun_post",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_task_id",
            "required": true,
            "schema": {
              "title": "Benchmark Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RerunBenchmarkTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "再次执行批量任务",
        "tags": [
          "基准任务"
        ]
      }
    },
    "/benchmark-tasks/{benchmark_task_id}/stop": {
      "post": {
        "operationId": "stop_benchmark_task_benchmark_tasks__benchmark_task_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_task_id",
            "required": true,
            "schema": {
              "title": "Benchmark Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "停止当前批量任务",
        "tags": [
          "基准任务"
        ]
      }
    },
...
  }
}